Fix running devicestest after requiring >= ECM 5.38
Requiring at least ECM 5.38 comes with a different build dir layout
to help testing of not installed apps. But in combination with the
old add_test syntax it caused qtvideosink_autotest to fail because
the test     executable wasn't found anymore. Using ecm_add_test
fixes that.

That being I said I'm not sure about the purpose of the test because
it doesn't test anything about Audex and should probably be removed.